delete property

+++ controllers/Propriety.js
@@ -41,7 +41,7 @@ const addPropriety = (async (req, res) => {
             totalUnits: req.body.totalUnits,
             occupiedUnits: req.body.occupiedUnits,
             availableUnits: req.body.availableUnits,
-            landLord : req.body.userId
+            landLord : req.userId
         })
         await propriety.save()
         console.log(propriety);
@@ -56,7 +56,7 @@ const addPropriety = (async (req, res) => {
         } */
         
         // const proprietyId = req.body.landlordNumber + '-' + req.body.proprietyName
-        const landlord = await Landlord.findOne({ landlordNumber: req.userId });
+        const landlord = await Landlord.findOne({ _id: req.userId });
         if(landlord){
             if (!landlord.listOfProprieties) {
                 landlord.listOfProprieties = [];
@@ -132,20 +132,22 @@ const getPropriety = (async (req, res) => {
 const deletePropriety = (async (req, res) => {
     try {
         const propriety = await Propriety.findById(req.params.id);
-        const elt = propriety.proprietyId;
+        if(!propriety){
+            return res.status(404).json({error : "Property not found"});
+        }
         const landlord = await Landlord.findById(propriety.landLord);
         if (!landlord) {
-            return res.status(404).send('no user find');
+            return res.status(404).json({error : "landLord not found"});
         }
         const proprieties = landlord.listOfProprieties
         const newProprieties = proprieties.filter(proprietyId => proprietyId !== req.params.id);
         landlord.listOfProprieties = newProprieties;
         await landlord.save();
         await Propriety.deleteOne({ _id: propriety._id.toString() })
-        res.send("Propriety correctly removed")
+        res.json({success : true, message : "Propriety correctly removed"});
     } catch (error) {
         console.log("error :"+error);
-        res.send("Propriety not correctly removed" + error);
+        res.json({ error: error.message});
     }
 })
